+++ sharingIsCaring/src/test/java/propra2/database/ProductTest.java
@@ -33,4 +33,31 @@ public class ProductTest {
         Assertions.assertThat(15.0).isEqualTo(response);
     }
 
+    @Test
+    public void allValuesSetFalse(){
+        Customer customer = new Customer();
+        Product product = new Product();
+        product.setId(10L);
+        product.setOwner(customer);
+        product.setAvailable(true);
+
+        Assert.assertEquals(false, product.allValuesSet());
+    }
+
+    @Test
+    public void allValuesSetTrue(){
+        Customer customer = new Customer();
+        Product product = new Product();
+        product.setId(10L);
+        product.setOwner(customer);
+        product.setAvailable(false);
+        product.setDeposit(4);
+        product.setDailyFee(123);
+        product.setBorrowedUntil(new java.util.Date());
+        product.setTitle("Putzfrau");
+        product.setDescription("Putzt alles blitzblank");
+
+        Assert.assertEquals(true, product.allValuesSet());
+    }
+
 }
